Rules fix
 * Conditions    Rules
 * -------------------------------------------
 * % 10 == 0		0
 * Under 50		Add 3
 * Over  50		Subtract 3	Can't go negative, assign 0
 * Over 150		1

Over 150 means 151 and greater, does not include 150.  Update fixes this issue.  However, the order of precedence of these rules is
reflected by the order of their execution.  This means, that the first rule has the highest precedence, the second rule has the second highest precedence and so on...  This means that sums/values evenly divisible by 10 are never compared because they are immediately tied to the first (highest precedence) rule.